backend/requirements.txt
Django>=5.0,<6.0
django-ninja>=1.0,<2.0
psycopg[binary]>=3.0
gunicorn==21.2.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
feedparser>=6.0.10
django-redis>=5.4.0
redis>=5.0.0
celery>=5.3.0
django-celery-beat>=2.5.0
requests>=2.31.0
trafilatura>=1.6.0
lxml>=5.1.0
